Trim down bindings, and fix compiler errors for new ggml version

+++ llama-rs/src/lib.rs
@@ -174,7 +174,7 @@ pub enum ModelKVMemoryType {
     /// 32-bit float.
     Float32,
 }
-impl From<ModelKVMemoryType> for i32 {
+impl From<ModelKVMemoryType> for u32 {
     fn from(value: ModelKVMemoryType) -> Self {
         match value {
             ModelKVMemoryType::Float16 => ggml::TYPE_F16,
@@ -883,11 +883,11 @@ impl Model {
                 let ftype = read_u32(&mut part_reader)?;
 
                 let mut nelements = 1;
-                let mut ne = [1i32, 1i32];
+                let mut ne = [1i64, 1i64];
 
                 #[allow(clippy::needless_range_loop)]
                 for i in 0..n_dims {
-                    ne[i] = read_i32(&mut part_reader)?;
+                    ne[i] = read_i32(&mut part_reader)? as i64;
                     nelements *= usize::try_from(ne[i])?;
                 }
 
@@ -954,7 +954,7 @@ impl Model {
                         });
                     }
                 } else if split_type == 0 {
-                    if tensor.get_ne()[0] / i32::try_from(n_parts)? != ne[0]
+                    if tensor.get_ne()[0] / i64::try_from(n_parts)? != ne[0]
                         || tensor.get_ne()[1] != ne[1]
                     {
                         return Err(LoadError::TensorWrongSize {
@@ -963,7 +963,7 @@ impl Model {
                         });
                     }
                 } else if tensor.get_ne()[0] != ne[0]
-                    || tensor.get_ne()[1] / i32::try_from(n_parts)? != ne[1]
+                    || tensor.get_ne()[1] / i64::try_from(n_parts)? != ne[1]
                 {
                     return Err(LoadError::TensorWrongSize {
                         tensor_name,
